+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Data-picker: Added the ability to toggle between range and exact date.
 
+
 ### Changed
 - Updated the top panel UX/UI
 
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GridGroup: 
   - incorrect size of the multi-line chart
   - fixed response when updating data sources
+- Data-picker: Fixed changing format
 
 ## [1.158.0]